vimrc better completion, remove ConvertFile, fixes
Combined TabComplete(), StepOver(), OpenBelow(), and
DeleteMatching() into a single Complete() function that can be toggled
on and off.

Removed the ConvertFile() since I never use it. Now the ctrl-c mapping
can be used to toggle completion.

Added commenting for *.less files.

Fixed a bug where the git branch would be wrong when switching buffers
between files in different git projects.

Moved the ApplyCustomTheme() logic out and removed the function. There's
no real reason why it needed to be a function.

Made it so that help pages are always full size in the window.