Use vim plugins. Add vundle, airline, and fugitive
I've finally decided to start using plugins. This is something i've
avoided doing for a long time but the vimrc is getting very large
as I've implemented a lot of IDE-like features over the years. There's a
lot of good plugins out there that are actively maintained and already
do what I've been trying to implement myself. Don't reinvent the wheel,
they say.

To start with I've just replaced the statusline with airline. Over time
I'll phase out my implementation of a bunch of things in favor of
plugins as well as add plugins for other cool stuff.