Remove easygrep
I didn't use it that much and it adds complexity to the [enter] mapping.
Plus there needed to be language-specific excludes like node_modules.
I'm not a fan of that. Maybe one day we can make use of ! and !! to
filter thru external program (grep) and use gf to open that file or
something like that.