Redefine keybindings in vimrc and tmux.conf
I want to avoid overriding default vim bindings as much as possible.
Instead of using ctrl-x, use alt for buffer and window actions, and use
mapleader for other shortcuts (which is was it is intended for).

Tmux now uses ctrl-alt instead of alt. Horizontal and vertical split now
use s and v since ctrl-alt-- doesn't seem to work right.