Fix some neovim term related issues
m-d now works to delete a buffer that comes after a terminal buffer.
Before, :bprevious would move to the term buffer, then the autocmd would
kick in to enter insert mode in the term and `:bdelete #` would be
executed in the shell. Now I'm using a pipe instead of two consecutive
commands.

Splitting a term window now makes a new terminal rather than showing the
same term in two windows.

Always just show the tail-end of a buffer name in airline's tabline.
Before, if there were two terminal buffers open, one of them would have
a long name (including the term:// part) and the other would just be
'bash'. Now they're all just 'bash'.
This means that if you have two files open with the same name,
their names will be indistinguishable in the tabline; that's ok, it's
pretty rare and not much of a problem anyway.