New keybindings for vim and tmux
Previously, vim used alt keybindings to navigate windows and tabs
because I wanted to be able to use them while in terminal mode in
neovim. Now just use <esc> to get out of terminal mode. I didn't
want to do this previously in case I accidentally ended up in insert
mode in a nested nvim. If this happens, <c-o> in insert mode can be used
to get out. Without any alt keybindings in vim, they can be used in
tmux now.

Now all custom ctrl keybindings in vim are navigational and map to the
same alt binding in tmux.