vimrc: use nerdtree instead of netrw (:Explore)
netrw is buggy and the copy/pasting is hard to use.
I initially tried to make a plugin that could bind c-c, c-v, etc. for
netrw but couldn't make it work; in netrw, you basically have to choose
where you want to paste to first, so c-v, then c-c. wtf. Use nerdtree