vimrc: add ctrlp, remove treestump
Used to use ctrlp but removed it because it was slow in large projects.
Using git ls-files makes it fast.

Treestump is alright but I didn't end up using it and it has problems
with copying a file to the same directory but with a new name.
nerdtree's menu (m) works really well.