bash_profile: don't source ~/.local_env
I noticed aliases defined in ~/.local_env weren't working in Gnome
because the profile is only executed on login, not when firing up a
terminal. Since ~/bin is added to the path, you could add a script in
there that sets up an environment and just run it when you need it.
For aliases, those can just be executables in ~/bin.