vimrc: ctrlp regex match instead of fuzzy
Was running into a problem where trying to find 'ab.js' was not actually
finding that file because fuzzy was matching stuff like Asdf/Bvcx.JS or
anything else that includes a and b somewhere in the path; as soon as 10
matches were found, the search stopped. With regex, the letters must be
consecutive.