vimrc: fix Glog mappings
The normal mode mapping wasn't working on files that weren't in the
project's root because the path wasn't relative to the --git-dir arg set
by Glog.

The visual mode mapping didn't work at all because the range was used
for the :silent command instead of Glog. Not sure how that got committed
without being tested...