bash_profile: use local_profile; no ~/bin in PATH
The path can be modified in local_profile and so can other env variables
like GOPATH or whatever.

It probably won't work to put aliases in local_profile but if you wanted
to set an alias for `some_cmd` you could add a bin dir first to the path
in local_profile (like the line that was removed in this commit), then
add a `some_cmd` file to ~/bin and have it call the real `some_cmd` like
`/usr/bin/some_cmd --some-alias-option $@`.