gitconfig: down alias uses stash message 'down'
When `git down` pops the stash and there's a conflict, it leaves the
stash in the list and I forget to drop it; then I come back after a
few days and look at my stash list and don't know if that's something
important or an old temporary `down` stash. Now those temporary stashes
will be labeled with 'down'.