Revert "vimrc: closing a tab selects the previous tab"
This reverts commit 8ca809a7a035511921aaefec632c274378a8b191.

This autocmd doesn't solve the problem it was meant to. When three tabs
are open and the last one is closed, this autocmd causes the first tab
to be selected, not the second. Until I can find a good way to deal with
this (or just decide to put up with the default behavior) I'm going to
pull this autocmd out.