vimrc: use chronolog instead of outlaw
Git grep and diff from the outlaw plugin are replaced by simpler
alternatives that can be done without a plugin. Plus chronolog provides
blame and show operations and is over half the size of outlaw.

Can't autochdir anymore because it causes git grep to work from the cwd
rather than the whole project. This also means that vim needs to be
started from the git project root.