vimrc: break up <space>i into separate commands
I found that i was using this often to see line numbers but didn't want
to have list or spelling turned on at the same time.

Turned off cursorline by default because it's kind of obnoxious and
hides spelling mistakes since my colorscheme marks bad spelling
underlined. The cursorline can now be easily be toggled if I need it.