plugs.vim: prevent :helptags ALL from failing
Noticed this started happening because /use/share/vim/vim80/doc/tags
could not be written to (write permission is only for root). Tho this
was working before... I'm guessing that the tags file is managed or
updated by the vim package installed on the machine so it got new
permissions?? Anyway, now we just generate help tags for any plugins
that include a doc/ folder rather than generating for every doc/ in the
runtime path.