Cut the fat; remove unused/unnecessary dotfiles
I never actually use gvim and the only reason I use neovim is because
the cursor changes to a bar in insert mode. If I really need a mode
indicator, I can set up an autocmd to toggle cursorline or something.
There's no sense in having multiple editors that do the same thing, and
vim in the common denominator.