vimrc: remove remaining plugin settings
It's hardly practical but it's clean. I don't want the dotfiles to have
dependencies or assume the presence of dependencies. Let them be
independent. I'm trying to boil them down to be as minimal as possible.
Plugin related settings can always be added locally and without needing
to be committed in git.

The formatoptions autocmd was added as a fix because of a setting in the
vim-javascript plugin that would set the option. I've actually been in
cases where I'm editing large blocks of comments and it's useful to have
it on, plus it's not hard to get around it by using O from a
non-commented line.