shortcuts/ledger: don't use readline for inputs
With readline, pressing tab will do file completion, which I don't need
for any of the inputs the script takes. I thought about making
completion work for tags but it seems the only way to make that work is
a hacky workaround by making a tmp dir and filling it with a file for
each tag name. Just type