vimrc: clipboard uses generic mappings
After installing Arch again I found that clipboard wasn't working for
some reason so I decided to change the way it works by using some
generic mappings that will also work with termux and also for testing
portions of scripts.

Unfortunately `:w !...` only works linewise so to work on a portion of a
line the best solution I could come up with was to first copy that
portion to its own line with another mapping. I don't like it but for
now I'll try it.