Replace dmenu_history with xmenu
This is mainly because dmenu will be replaced by wmenu, which is now the
sway default. Because of this, I won't be using dmenu_run anymore, which
is totally fine because I don't need a list of every single executable
every time I press ctrl-d. This will also perform better and avoid the
edge case where there's duplicate entries.

xmenu takes the program as an argument like xargs so it's flexible to
work with dmenu, bemenu, wmenu, etc. It also passes -l to the program
since it knows how many lines there will be.