Use default colors
I did this mainly because I find myself often distracted by making
adjustments to colors and themes. Everything - vim, sway, foot, termux,
etc. defaults to a dark theme so it works out to just leave it that way.
Plus now everything is much simpler and closer to default.

I chose slate as vim's colorscheme because I think it looks the best out
of the included ones.

foot cursor isn't reversed because it gets hard to see when on matched
parentheses.

local/bin/colortest was moved to github.com/tylerbrazier/archive