foot.ini: term=xterm-256color
For some reason I never had a problem with ssh into termux on Arch
but in Alpine I'm seeing color problems and errors when starting vim.
I didn't see any kind of foot terminfo package for termux either so I
guess this is the best fix; and it would be good to keep for remoting
into any other kind of server anyway.